Revert "refactor: UploadComponent"
This reverts commit 247c023212dcce65d5cbf9bec46adaf43118bf5c.
