+++ src/app/post/[...slug]/page.tsx
@@ -1,13 +1,13 @@
 import CommentsComponent from "@/app/dao/components/comments"
 import { MarkdownRenderers } from "@/app/upload/utils/MarkdownRenderers"
-import AuthorAvatar from "@/components/AuthorAvatar"
 import HiveClient from "@/lib/hive/hiveclient"
 import { transform3SpeakContent, transformIPFSContent } from "@/lib/utils"
-import { Badge, Box, Center, Container, Divider, HStack, Heading, Table, Tbody, Td, Text, Th, Thead, Tr, VStack } from "@chakra-ui/react"
+import { Avatar, Badge, Box, Center, Container, Divider, HStack, Heading, Table, Tbody, Td, Text, Th, Thead, Tr, VStack } from "@chakra-ui/react"
 import { Metadata } from "next"
 import ReactMarkdown from "react-markdown"
 import rehypeRaw from "rehype-raw"
 import remarkGfm from "remark-gfm"
+import AuthorAvatar from "@/components/AuthorAvatar"
 // Revalidate requests in 10 minutes
 export const revalidate = 600
