+++ src/app/upload/components/previewModal.tsx
@@ -1,31 +1,38 @@
+// path: src/app/upload/components/previewModal.tsx
+
+import Footer from '@/components/PostCard/Footer';
+import Header from '@/components/PostCard/Header';
+import PostImage from '@/components/PostCard/Image';
+import { PostProvider } from '@/contexts/PostContext';
 import { commentWithKeychain } from '@/lib/hive/client-functions';
-import { commentWithPrivateKey } from '@/lib/hive/server-functions';
+import { commentWithPrivateKey } from "@/lib/hive/server-functions";
 import { HiveAccount } from '@/lib/useHiveAuth';
 import {
+    Badge,
     Box,
+    Button,
+    Card,
+    CardHeader,
     Center,
     Divider,
-    Modal,
-    ModalBody,
-    ModalCloseButton,
-    ModalContent,
-    ModalFooter,
-    ModalHeader,
-    ModalOverlay,
-    VStack
+    Flex,
+    Modal, ModalBody, ModalCloseButton, ModalContent, ModalFooter, ModalHeader, ModalOverlay,
+    Progress,
+    Table,
+    Tbody,
+    Td,
+    Text,
+    Th,
+    Thead,
+    Tr,
+    VStack,
 } from '@chakra-ui/react';
-import React, { useState } from 'react';
-
+import * as dhive from "@hiveio/dhive";
+import React, { useEffect } from 'react';
+import getSummary from '../../../lib/getSummaryAI';
+import generatePermlink from '../utils/generatePermlink';
 import slugify from '../utils/slugify';
-import BeneficiariesTable from './BeneficiariesTable';
-import { createCommentOptions, createPostOperation, formatBeneficiaries } from './formatPostData';
-import ModalFooterButtons from './ModalFooterButtons';
-import PostPreview from './PostPreview';
 import SocialsModal from './socialsModal';
-import TagsTable from './TagsTable';
-import usePostLink from './usePostLink';
-import useSummary from './useSummary';
-
 interface PreviewModalProps {
     isOpen: boolean;
     onClose: () => void;
@@ -36,11 +43,42 @@ interface PreviewModalProps {
     beneficiariesArray: any[];
     tags: string[];
 }
+interface BeneficiaryForBroadcast {
+    account: string;
+    weight: string;
+}
+interface BeneficiariesCard {
+    beneficiariesArray: BeneficiaryForBroadcast[];
+}
+const BeneficiariesCard: React.FC<BeneficiariesCard> = ({ beneficiariesArray }) => {
+    return (
+        <Card bg='darkseagreen' border="1px solid limegreen">
+            <CardHeader>
+                <VStack spacing={4} align="stretch">
+                    <Box>
+                        <Text color={"black"}>
+                            You - {100 - beneficiariesArray.reduce((acc, cur) => acc + Number(cur.weight), 0) / 100}%
+                        </Text>
+                        <Progress colorScheme="green" size="sm" value={100 - beneficiariesArray.reduce((acc, cur) => acc + Number(cur.weight), 0) / 100} />
+                    </Box>
+                    {beneficiariesArray.map((beneficiary, index) => (
+                        <Box key={index}>
+                            <Text color="black">
+                                {beneficiary.account} - {Number(beneficiary.weight) / 100}%
+                            </Text>
+                            <Progress colorScheme="green" size="sm" value={Number(beneficiary.weight) / 100} />
+                        </Box>
+                    ))}
+                </VStack>
+            </CardHeader>
+        </Card>
+    );
+};
 
 const PreviewModal: React.FC<PreviewModalProps> = ({ isOpen, onClose, title, body, thumbnailUrl, user, beneficiariesArray, tags }) => {
-    const [hasPosted, setHasPosted] = useState(false);
-    const AiSummary = useSummary(body);
-    const postLink = usePostLink(title, user);
+    const [hasPosted, setHasPosted] = React.useState(false);
+    const [postLink, setPostLink] = React.useState("");
+    const [AiSummary, setAiSummary] = React.useState("");
 
     let postDataForPreview = {
         post_id: Number(1),
@@ -61,11 +99,39 @@ const PreviewModal: React.FC<PreviewModalProps> = ({ isOpen, onClose, title, bod
             { voter: "Magnolia", weight: 20000, percent: "0", reputation: 100, rshares: 0 },
         ]
     }
+    useEffect(() => {
+        const summarizeBlog = async () => {
+            const summary = await getSummary(body);
+            setAiSummary(summary);
+        };
+
+        if (body) {
+            summarizeBlog();
+        }
+    }, [body]);
 
     const handlePost = async () => {
+        const formatBeneficiaries = (beneficiariesArray: any[]) => {
+            let seen = new Set();
+            return beneficiariesArray.filter(({ account }: { account: string }) => {
+                const duplicate = seen.has(account);
+                seen.add(account);
+                return !duplicate;
+            }).map(beneficiary => ({
+                account: beneficiary.account,
+                weight: parseInt(beneficiary.weight, 10) // Ensure weight is an integer
+            })).sort((a, b) => a.account.localeCompare(b.account));
+        };
+
         const finalBeneficiaries = formatBeneficiaries(beneficiariesArray);
         const permlink = slugify(title.toLowerCase());
         const loginMethod = localStorage.getItem('LoginMethod');
+
+        const summarizeBlog = async () => {
+            const summary = await getSummary(body);
+            setAiSummary(summary);
+        };
+
 
         if (!user || !title || !user.name) {
             if (loginMethod === 'keychain') {
@@ -106,12 +172,56 @@ const PreviewModal: React.FC<PreviewModalProps> = ({ isOpen, onClose, title, bod
                 alert('Something went wrong, please try again');
             }
         } else if (loginMethod === 'privateKey') {
-            const commentOptions = createCommentOptions(user, permlink, finalBeneficiaries);
-            const postOperation = createPostOperation(user, permlink, title, body, tags, thumbnailUrl);
+            const commentOptions: dhive.CommentOptionsOperation = [
+                'comment_options',
+                {
+                    author: user.name,
+                    permlink: permlink,
+                    max_accepted_payout: '10000.000 HBD',
+                    percent_hbd: 10000,
+                    allow_votes: true,
+                    allow_curation_rewards: true,
+                    extensions: [
+                        [0, {
+                            beneficiaries: finalBeneficiaries
+                        }]
+                    ]
+                }
+            ];
+
+            const postOperation: dhive.CommentOperation = [
+                'comment',
+                {
+                    parent_author: '',
+                    parent_permlink: 'hive-173115',
+                    author: user.name,
+                    permlink: permlink,
+                    title: title,
+                    body: body,
+                    json_metadata: JSON.stringify({
+                        tags: tags,
+                        app: 'Skatehive App',
+                        image: thumbnailUrl,
+                    }),
+                },
+            ];
+
             commentWithPrivateKey(localStorage.getItem("EncPrivateKey"), postOperation, commentOptions);
             setHasPosted(true);
         }
     }
+
+    useEffect(() => {
+        const buildPostLink = () => {
+            const username = user?.name;
+            if (username) {
+                const permlink = generatePermlink(title);
+                const link = `https://skatehive.app/post/hive-173115/@${username}/${permlink}`;
+                setPostLink(link);
+            }
+        }
+        buildPostLink();
+    }, [title, user]);
 
     return (
         <Modal isOpen={isOpen} onClose={onClose} size="xl">
@@ -129,18 +239,58 @@ const PreviewModal: React.FC<PreviewModalProps> = ({ isOpen, onClose, title, bod
                                 <Center>
                                     <VStack>
                                         <Box width={"100%"}>
-                                            <PostPreview postData={postDataForPreview} />
+                                            <Card
+                                                bg={"black"}
+                                                border={"0.6px solid white"}
+                                                size="sm"
+                                                boxShadow="none"
+                                                borderRadius="none"
+                                                p={2}
+                                            >
+                                                <PostProvider postData={postDataForPreview}>
+                                                    <Header />
+                                                    <PostImage />
+                                                    <Footer />
+                                                </PostProvider>
+                                            </Card>
                                         </Box>
                                         <Box border={'1px solid white'} w="sm" maxWidth="sm">
-                                            <BeneficiariesTable beneficiariesArray={beneficiariesArray} />
-                                            <TagsTable tags={tags} />
+                                            <Table>
+                                                <Thead>
+                                                    <Tr>
+                                                        <Th>Beneficiaries</Th>
+                                                    </Tr>
+                                                </Thead>
+                                                <Tbody>
+                                                    <Tr>
+                                                        <Td>
+                                                            <BeneficiariesCard beneficiariesArray={beneficiariesArray} />
+                                                        </Td>
+                                                    </Tr>
+                                                </Tbody>
+                                            </Table>
+                                            <Table>
+                                                <Th>Tags</Th>
+                                                <Td>
+                                                    <Flex flexWrap="wrap">
+                                                        {tags.map((tag, index) => (
+                                                            <Badge key={index} colorScheme="green" variant="solid" size="sm" m={1}>
+                                                                {tag}
+                                                            </Badge>
+                                                        ))}
+                                                    </Flex>
+                                                </Td>
+                                            </Table>
                                         </Box>
                                     </VStack>
                                 </Center>
                             </Box>
                         </ModalBody>
                         <ModalFooter>
-                            <ModalFooterButtons onClose={onClose} handlePost={handlePost} />
+                            <VStack width={'100%'}>
+                                <Button width={'100%'} colorScheme='red' onClick={onClose}>Let me try again, I am high</Button>
+                                <Button width={'100%'} colorScheme='green' onClick={handlePost}>Looks dope, confirm!</Button>
+                            </VStack>
                         </ModalFooter>
                     </>
                 )}
